Feature/iso 1691 verificar se os stop ids no ficheiro stopstxt existem (#56)
* feat(stops): validate stop_id against stops_ids.json reference list

Pre-compute valid stop IDs from root-level stops_ids.json at validation
startup and reject stop_ids not present in the reference set.

* feat(stops): validate stops against stops_data.json reference

Replace stops_ids.json ID-only lookups with a stops_data cache that
indexes name, latitude, and longitude by stop_id. Cross-check stop_id,
stop_name, stop_lat, and stop_lon against the reference data.

* Merge branch 'production' into feature/iso-1691-verificar-se-os-stop_ids-no-ficheiro-stopstxt-existem

* feat(stops): enhance validation rules and error messages for stop data

- Added new validation rules for stop_id_exists, stop_name_matches_stops_data, stop_lat_matches_stops_data, and stop_lon_matches_stops_data in tml-rules.json.
- Updated error messages in English and Portuguese for stop_name, stop_lat, and stop_lon validations to include checks against stops_data.json.
- Refactored validation functions to incorporate new rules and improve error reporting consistency across stop validations.

* fix(validation): update stop name validation rule identifier for consistency

- Changed the identifier for the stop name validation rule from "stops.stop_name_matches_data_cache" to "stops.stop_name_matches_stops_data" in both the rules parser and GTFS rules definitions for improved clarity and consistency.

* refactor(validation): update GTFS file handling and error reporting

- Normalized file names in validation messages to handle nested ZIP structures consistently.
- Updated error logging to reflect the base name of files instead of full paths for clarity.
- Enhanced validation messages for unsupported files, ensuring they are ignored appropriately.
- Improved handling of required GTFS files, adding error messages for missing mandatory files.

* refactor(validation): adjust severity levels for stop validation rules

- Changed severity levels for the following validation rules from "error" to "warning": stop_name_matches_stops_data, stop_lat_matches_stops_data, and stop_lon_matches_stops_data. This adjustment aims to reduce the strictness of these validations while still highlighting potential issues.

* Update validator/types/cache.go


* refactor(validation): update latitude and longitude types from float32 to float64

- Changed the data types for latitude and longitude in the Stop, Shape, and validation functions to float64 for improved precision.
- Updated related validation tests and parsing logic to accommodate the new data types, ensuring consistency across the validation framework.

* refactor(validation): streamline stop name validation logic

- Reorganized the stop name validation function to improve readability and maintainability.
- Ensured that the validation checks for stop name options are correctly structured, enhancing the clarity of the validation flow.

* Update validator/validations/stops/precompute.go


* fix(validation): correct typo for non-existent stop_id

* Enhance stops validation by integrating stops_data.json support

- Updated CLI options to include a path for stops_data.json.
- Modified the BuildStopsDataCache function to read from the specified stops_data.json path, improving flexibility in data handling.
- Added error handling for stops data cache initialization and logging for better debugging.
- Adjusted Makefile to reflect the new stops_data.json dependency in the development run command.

* Enhance latitude and longitude validations with tolerance options

- Added tolerance options for stop latitude and longitude validations in tml-rules.json, allowing for more flexible matching against data.
- Updated StopLatValidation and StopLonValidation functions to incorporate tolerance checks, improving validation accuracy while accommodating minor discrepancies.

* Enhance StopsDataCache structure and validation logic

- Added ValidStopIDs field to StopsDataCache for improved tracking of valid stop_ids from flags.
- Updated BuildStopsDataCache function to initialize ValidStopIDs, ensuring it captures valid stop_ids during cache creation.
- Modified StopIdValidation to check against ValidStopIDs instead of ByStopID, refining the validation process for stop_ids.
- Expanded test cases to validate the new logic for handling internal IDs and valid stop_ids.

* Remove stops_data.json file from the repository, eliminating outdated stop information and associated flags. This change streamlines the data structure and prepares for future updates to stop data management.

* Refactor stop validation functions to remove unused parameters

- Updated StopIdValidation, StopLatValidation, StopLonValidation, and StopNameValidation functions to eliminate the stopsData parameter, simplifying the validation logic.
- Adjusted corresponding test cases to reflect the changes in function signatures, ensuring consistency across validations.
- Commented out sections of code related to stopsData checks for potential future use, maintaining clarity in the validation process.

* Update Makefile to remove stops_data.json dependency from dev command

- Modified the development run command in the Makefile to eliminate the stops_data.json argument, simplifying the command and reflecting recent changes in data management.

* Refactor stop validation rules to remove deprecated checks

- Removed unused validation rules for stop_id_exists, stop_name_matches_stops_data, stop_lat_matches_stops_data, and stop_lon_matches_stops_data from tml-rules.json and corresponding validation logic in rules_parser.go and gtfs_rules.go.
- This cleanup simplifies the validation process and aligns with recent changes in data management.

* Merge branch 'production' into feature/iso-1691-verificar-se-os-stop_ids-no-ficheiro-stopstxt-existem

* Refactor ParseStop function to enhance clarity and consistency in variable declarations. Updated the stop access variable to improve validation logic and maintainability.

+++ validator/lib/validators.go
@@ -95,11 +95,11 @@ func ValidatePhone(p string) bool {
 	return len(digitMatches) >= 7
 }
 
-func ValidateLongitude(lon float32) bool {
+func ValidateLongitude(lon float64) bool {
 	return lon >= -180 && lon <= 180
 }
 
-func ValidateLatitude(lat float32) bool {
+func ValidateLatitude(lat float64) bool {
 	return lat >= -90 && lat <= 90
 }
